# util/buffer_pool, db: Reduce memory allocation