--- a/leveldb/table/writer.go
+++ b/leveldb/table/writer.go
@@ -146,6 +146,7 @@ type Writer struct {
 	compression opt.Compression
 	blockSize   int
 
+	bpool       *util.BufferPool
 	dataBlock   blockWriter
 	indexBlock  blockWriter
 	filterBlock filterWriter
@@ -285,6 +286,16 @@ func (w *Writer) BytesLen() int {
 // after Close, but calling BlocksLen, EntriesLen and BytesLen
 // is still possible.
 func (w *Writer) Close() error {
+	defer func() {
+		if w.bpool != nil {
+			// Buffer.Bytes() returns [offset:] of the buffer.
+			// We need to Reset() so that the offset = 0, resulting
+			// in buf.Bytes() returning the whole allocated bytes.
+			w.dataBlock.buf.Reset()
+			w.bpool.Put(w.dataBlock.buf.Bytes())
+		}
+	}()
+
 	if w.err != nil {
 		return w.err
 	}
@@ -351,7 +362,15 @@ func (w *Writer) Close() error {
 // NewWriter creates a new initialized table writer for the file.
 //
 // Table writer is not safe for concurrent use.
-func NewWriter(f io.Writer, o *opt.Options) *Writer {
+func NewWriter(f io.Writer, o *opt.Options, pool *util.BufferPool, size int) *Writer {
+	var bufBytes []byte
+	if pool == nil {
+		bufBytes = make([]byte, size)
+	} else {
+		bufBytes = pool.Get(size)
+	}
+	bufBytes = bufBytes[:0]
+
 	w := &Writer{
 		writer:          f,
 		cmp:             o.GetComparer(),
@@ -359,6 +378,8 @@ func NewWriter(f io.Writer, o *opt.Options) *Writer {
 		compression:     o.GetCompression(),
 		blockSize:       o.GetBlockSize(),
 		comparerScratch: make([]byte, 0),
+		bpool:           pool,
+		dataBlock:       blockWriter{buf: *util.NewBuffer(bufBytes)},
 	}
 	// data block
 	w.dataBlock.restartInterval = o.GetBlockRestartInterval()